[cocas] Reorganize code of object_module
