+++ cocas/object_file/__init__.py
@@ -1,1 +1,2 @@
-from .object_file import import_object
+from .object_export import export_object
+from .object_import import import_object
