+++ cocas/object_module/__init__.py
@@ -1,3 +1,3 @@
 from .external_entry import ExternalEntry
 from .location import CodeLocation
-from .object_module import ObjectModule, ObjectSectionRecord, export_objects
+from .object_module import ObjectModule, ObjectSectionRecord
